Pushed extra note about handling ediff in the same frame.
The origin for this comes from the fact that the ediff frame behaves
weirdly on my xmonad, and I really wanted a fix. This makes it far
easier to use ediff without size-of-line issues within the frame.

The way this works is to embed the control window into the bottom
between the two buffers being diffed, and seems to discard all the
pretty frame handling of the ediff control frame, putting it into
the control window instead.